v 4

+++ database/migrations/2021_09_16_140545_create_partidas_table.php
@@ -15,7 +15,7 @@ class CreatePartidasTable extends Migration
     {
         Schema::create('partidas', function (Blueprint $table) {
             $table->increments('id_partida');
-            $table->double('activa');
+            $table->integer('activa');
             $table->integer('max_usuarios');
             $table->timestamps();
